Fix a whole lot of conversion warnings
A lot of these numeric_casts are there solely to prevent corruption in
case we do overflow.

+++ src/vw/FileIO/DiskImageResourcePNG.cc
@@ -612,7 +612,7 @@ void DiskImageResourcePNG::read_reset() const {
 
 unsigned DiskImageResourcePNG::num_comments() const {
   m_ctx->read_comments();
-  return m_ctx->comments.size();
+  return boost::numeric_cast<unsigned>(m_ctx->comments.size());
 }
 
 DiskImageResourcePNG::Comment const& DiskImageResourcePNG::get_comment( unsigned i ) const {
